Many changes, see description
Added classes inputs and outputs to control and display the logic of the other gates. Added a drag threshold to control whether inputs are toggling or being moved. Added connection points to wire gates up. Added logic to draw the connection wires. Added a function to recalc the entire circuit to dynamically show changes (that one's not working yet - but it's in there!)